Extract_text_from_pdfs/requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF==1.22.5
PyPDF2==3.0.1